requirements.txt
altair==4.2.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
distro==1.9.0
entrypoints==0.4
exceptiongroup==1.2.2
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib_metadata==8.6.1
Jinja2==3.1.6
jiter==0.9.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
narwhals==1.33.0
numpy==2.0.2
openai==1.70.0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
protobuf==3.20.3
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyarrow==19.0.1
pydantic==2.11.2
pydantic_core==2.33.1
pydeck==0.9.1
Pygments==2.19.1
Pympler==1.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pywin32==310
referencing==0.36.2
requests==2.32.3
rich==14.0.0
rpds-py==0.24.0
semver==3.0.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.40
streamlit==1.12.0
toml==0.10.2
toolz==1.0.0
tornado==6.4.2
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.3.0
validators==0.34.0
watchdog==6.0.0
zipp==3.21.0
